requirements.txt
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
outcome==1.3.0.post0
pip==23.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PySocks==1.7.1
pyTelegramBotAPI==4.20.0
python-telegram-bot==21.3
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
telebot==0.0.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
